location auto-detecting

+++ app/src/main/java/ru/glindaquint/weatherapp/types/openWeatherMap/Wind.kt
@@ -3,6 +3,6 @@ package ru.glindaquint.weatherapp.types.openWeatherMap
 import com.google.gson.annotations.SerializedName
 
 data class Wind(
-    @SerializedName("speed") var speed: Int? = null,
-    @SerializedName("deg") var deg: Int? = null,
+    @SerializedName("speed") var speed: Double? = null,
+    @SerializedName("deg") var deg: Double? = null,
 )
